+++ src/ts/data/Tag.ts
@@ -1,4 +1,5 @@
 import { serverValidTags } from "@/ts/servers/CurrentServerRedirect";
+import { sanitize } from "@/ts/utils/TextUtils";
 
 export class Tag {
     public type: string;
@@ -55,5 +56,34 @@ export class TagNode {
                 return tag;
         }
         return undefined;
+    }
+
+    public static newFromSearch(search: string) {
+        let tagNode = new TagNode();
+
+        if (!search.includes(":")) {
+            tagNode.addRemaining(sanitize(search));
+            return tagNode;
+        }
+
+        const parts = search.split(" ");
+
+        for (const part of parts) {
+            if (!part.includes(":")) {
+                tagNode.addRemaining(part);
+                continue;
+            }
+            const splittedTag = part.split(":");
+            tagNode.addTag(
+                new Tag(sanitize(splittedTag[0]), sanitize(splittedTag[1]))
+            );
+        }
+
+        // tagNode.getTags().forEach(element => {
+        //     console.log(element.type.toString() + " of val " + element.value);
+        // });
+        // console.log("remaining: " + tagNode.getRemaining());
+        
+        return tagNode;
     }
 }
