+++ src/ts/manager/CurrentMap.ts
