--- src/ts/manager/MapsManager.ts
@@ -1,1 +1,0 @@
-// TODO: move SearchEngine's map saving logic here.
